move channel switcher back to learn
