+++ kolibri/plugins/learn/assets/src/views/channel-switcher/index.vue
